move ModelCard to components folder
